split fv/ffs code into separate file
